New Tab Page: Move sponsored images component to correct browser subdirectory
