+++ browser/ntp_sponsored_images/new_tab_page_branded_view_counter.cc
@@ -16,8 +16,8 @@
 #include "brave/components/brave_ads/browser/ads_service_factory.h"
 #include "brave/components/brave_rewards/browser/rewards_service.h"
 #include "brave/components/brave_rewards/browser/rewards_service_factory.h"
-#include "brave/components/ntp_sponsored_images/ntp_sponsored_images_data.h"
-#include "brave/components/ntp_sponsored_images/url_constants.h"
+#include "brave/components/ntp_sponsored_images/browser/ntp_sponsored_images_data.h"
+#include "brave/components/ntp_sponsored_images/browser/url_constants.h"
 #include "chrome/browser/profiles/incognito_helpers.h"
 #include "chrome/browser/profiles/profile.h"
 #include "chrome/common/chrome_features.h"
