+++ components/ntp_sponsored_images/browser/ntp_sponsored_images_data.cc
@@ -3,10 +3,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/ntp_sponsored_images/ntp_sponsored_images_data.h"
+#include "brave/components/ntp_sponsored_images/browser/ntp_sponsored_images_data.h"
 
 #include "base/logging.h"
-#include "brave/components/ntp_sponsored_images/ntp_sponsored_images_internal_data.h"
+#include "brave/components/ntp_sponsored_images/browser/ntp_sponsored_images_internal_data.h"
 
 NTPSponsoredImagesData::NTPSponsoredImagesData() = default;
