+++ components/ntp_sponsored_images/browser/ntp_sponsored_images_data.h
