+++ components/ntp_sponsored_images/browser/ntp_sponsored_images_service.cc
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/ntp_sponsored_images/ntp_sponsored_images_service.h"
+#include "brave/components/ntp_sponsored_images/browser/ntp_sponsored_images_service.h"
 
 NTPSponsoredImagesService::NTPSponsoredImagesService(
     content::BrowserContext* browser_context,
