+++ components/ntp_sponsored_images/browser/ntp_sponsored_images_service.h
@@ -6,7 +6,7 @@
 #ifndef BRAVE_COMPONENTS_NTP_SPONSORED_IMAGES_NTP_SPONSORED_IMAGES_SERVICE_H_
 #define BRAVE_COMPONENTS_NTP_SPONSORED_IMAGES_NTP_SPONSORED_IMAGES_SERVICE_H_
 
-#include "brave/components/ntp_sponsored_images/ntp_sponsored_images_component_manager.h"
+#include "brave/components/ntp_sponsored_images/browser/ntp_sponsored_images_component_manager.h"
 #include "components/keyed_service/core/keyed_service.h"
 
 namespace content {
