+++ components/ntp_sponsored_images/browser/regional_component_data.h
