+++ components/ntp_sponsored_images/browser/switches.cc
@@ -3,7 +3,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // you can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "brave/components/ntp_sponsored_images/switches.h"
+#include "brave/components/ntp_sponsored_images/browser/switches.h"
 
 #include "base/command_line.h"
