+++ components/ntp_sponsored_images/browser/url_constants.h
